TENTATIVE-data/requirements.txt
flask
requests
mysql-connector-python
